Upgrade to the v5 myqdevice.com API
Fixes caseywebdev/homebridge-chamberlain#54

Based on research from:
[1] @dxdc https://github.com/luisiam/homebridge-liftmaster2/issues/63#issuecomment-544077034
[2] @chanomie https://github.com/chanomie/homebridge-myq#open-the-garage-door

I'm running with this now, and it seems to correctly monitor the state
of the door, and I can open and close it from the Home app.

I left the code structure mostly the same, but there are probably some
opportunities for improvements as the v5 API is quite different from the
previous one.