Additional changes for ECONNRESET issue
Since submitting previous updates I started receiving additional but different ECONNRESET errors. Tracing back to the api.js code I changed "/api/v5/Accounts" to "/api/v5.1/Accounts" in 2 places as well as in the headers: section. 

Additionally updated npm to @latest (v6.14.3). Have not observed any ECONNRESET errors in the last 2 days from any of my plugins which may indicate that the change to npm provides a general fix for the problem. Reviewed the nom release notes and not obvious that there were specific changes that would address this issue as a general fix.